frontend: k8s: Move cluster hooks to api/v1/hooks to break circular import
